Move non functional test to integration tests
